// app/services/fetch-api/models/CodecType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jellyfin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum CodecType {
    Video = 'Video',
    VideoAudio = 'VideoAudio',
    Audio = 'Audio'
}

export function CodecTypeFromJSON(json: any): CodecType {
    return CodecTypeFromJSONTyped(json, false);
}

export function CodecTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): CodecType {
    return json as CodecType;
}

export function CodecTypeToJSON(value?: CodecType | null): any {
    return value as any;
}