// app/services/fetch-api/models/DlnaProfileType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jellyfin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum DlnaProfileType {
    Audio = 'Audio',
    Video = 'Video',
    Photo = 'Photo'
}

export function DlnaProfileTypeFromJSON(json: any): DlnaProfileType {
    return DlnaProfileTypeFromJSONTyped(json, false);
}

export function DlnaProfileTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): DlnaProfileType {
    return json as DlnaProfileType;
}

export function DlnaProfileTypeToJSON(value?: DlnaProfileType | null): any {
    return value as any;
}