// app/services/fetch-api/models/EncodingContext.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jellyfin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum EncodingContext {
    Streaming = 'Streaming',
    Static = 'Static'
}

export function EncodingContextFromJSON(json: any): EncodingContext {
    return EncodingContextFromJSONTyped(json, false);
}

export function EncodingContextFromJSONTyped(json: any, ignoreDiscriminator: boolean): EncodingContext {
    return json as EncodingContext;
}

export function EncodingContextToJSON(value?: EncodingContext | null): any {
    return value as any;
}