// app/services/fetch-api/models/ProfileConditionType.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jellyfin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum ProfileConditionType {
    Equals = 'Equals',
    NotEquals = 'NotEquals',
    LessThanEqual = 'LessThanEqual',
    GreaterThanEqual = 'GreaterThanEqual',
    EqualsAny = 'EqualsAny'
}

export function ProfileConditionTypeFromJSON(json: any): ProfileConditionType {
    return ProfileConditionTypeFromJSONTyped(json, false);
}

export function ProfileConditionTypeFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfileConditionType {
    return json as ProfileConditionType;
}

export function ProfileConditionTypeToJSON(value?: ProfileConditionType | null): any {
    return value as any;
}