// app/services/fetch-api/models/ProfileConditionValue.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jellyfin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum ProfileConditionValue {
    AudioChannels = 'AudioChannels',
    AudioBitrate = 'AudioBitrate',
    AudioProfile = 'AudioProfile',
    Width = 'Width',
    Height = 'Height',
    Has64BitOffsets = 'Has64BitOffsets',
    PacketLength = 'PacketLength',
    VideoBitDepth = 'VideoBitDepth',
    VideoBitrate = 'VideoBitrate',
    VideoFramerate = 'VideoFramerate',
    VideoLevel = 'VideoLevel',
    VideoProfile = 'VideoProfile',
    VideoTimestamp = 'VideoTimestamp',
    IsAnamorphic = 'IsAnamorphic',
    RefFrames = 'RefFrames',
    NumAudioStreams = 'NumAudioStreams',
    NumVideoStreams = 'NumVideoStreams',
    IsSecondaryAudio = 'IsSecondaryAudio',
    VideoCodecTag = 'VideoCodecTag',
    IsAvc = 'IsAvc',
    IsInterlaced = 'IsInterlaced',
    AudioSampleRate = 'AudioSampleRate',
    AudioBitDepth = 'AudioBitDepth'
}

export function ProfileConditionValueFromJSON(json: any): ProfileConditionValue {
    return ProfileConditionValueFromJSONTyped(json, false);
}

export function ProfileConditionValueFromJSONTyped(json: any, ignoreDiscriminator: boolean): ProfileConditionValue {
    return json as ProfileConditionValue;
}

export function ProfileConditionValueToJSON(value?: ProfileConditionValue | null): any {
    return value as any;
}