// app/services/fetch-api/models/SubtitleDeliveryMethod.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Jellyfin API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * 
 * @export
 * @enum {string}
 */
export enum SubtitleDeliveryMethod {
    Encode = 'Encode',
    Embed = 'Embed',
    External = 'External',
    Hls = 'Hls'
}

export function SubtitleDeliveryMethodFromJSON(json: any): SubtitleDeliveryMethod {
    return SubtitleDeliveryMethodFromJSONTyped(json, false);
}

export function SubtitleDeliveryMethodFromJSONTyped(json: any, ignoreDiscriminator: boolean): SubtitleDeliveryMethod {
    return json as SubtitleDeliveryMethod;
}

export function SubtitleDeliveryMethodToJSON(value?: SubtitleDeliveryMethod | null): any {
    return value as any;
}